refactor(ui): split App.tsx orchestration and dashboard
Drop App.tsx from 1178 to 773 lines, clearing the 800-line red-flag
budget. Four sibling files now own pieces previously inlined in App.

- app/appInterfaceHelpers.ts (123): profile interface types plus
  findBestInterface / interfaceExistsWithType / parseProfileInterfaces
  / applyInterfaceRestoration helpers
- hooks/useDeviceScan.ts (102): triggerDeviceScan callback plus the
  poll/timeout refs and unmount cleanup that used to sit in App
- components/app/AppFooter.tsx (103): product info, contact,
  website, legal links, and copyright
- components/app/AppDashboard.tsx (197): the four dashboard sections
  (connectivity, network, testing & discovery, system)

Pure relocation - behavior and public surface unchanged. The unused
biome-ignore on App's complexity suppression dropped naturally once
the dashboard JSX moved.
